added a admin route with different components to the pages
