+++ src/components/AdminHome/AdminHome.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+// This is one of our simplest components
+// It doesn't have local state,
+// It doesn't dispatch any redux actions or display any part of redux state
+// or even care what the redux state is'
+
+function AdminHome() {
+  return (
+    <div className="container">
+      <div>
+        <p>This is the admin page!</p>
+      </div>
+    </div>
+  );
+}
+
+export default AdminHome;
