+++ src/components/App/App.jsx
@@ -13,29 +13,15 @@ import Footer from '../Footer/Footer';
 
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
 
+import Admin from '../Admin/Admin';
 import AboutPage from '../AboutPage/AboutPage';
 import UserPage from '../UserPage/UserPage';
 import InfoPage from '../InfoPage/InfoPage';
 import LandingPage from '../LandingPage/LandingPage';
-import LoginPage from '../LoginPage/LoginPage';
-import RegisterPage from '../RegisterPage/RegisterPage';
-
-
-import Admin from '../Admin/Admin';
-import AdminLanguage from '../AdminLanguage/AdminLanguage';
-import AdminCategory from '../AdminCategory/AdminCategory';
-import AdminApprovals from '../AdminApprovals/AdminApprovals';
 
 import './App.css';
 
 function App() {
-  const dispatch = useDispatch();
-
-  const user = useSelector(store => store.user);
-
-  useEffect(() => {
-    dispatch({ type: 'FETCH_USER' });
-  }, [dispatch]);
 
   return (
     <Router>
@@ -57,62 +43,18 @@ function App() {
             Visiting localhost:3000/user will show the UserPage if the user is logged in.
             If the user is not logged in, the ProtectedRoute will show the LoginPage (component).
             Even though it seems like they are different pages, the user is always on localhost:3000/user */}
-          <ProtectedRoute
-            // logged in shows UserPage else shows LoginPage
-            exact
-            path="/admin"
-          >
-            <Admin />
-          </ProtectedRoute>
-
-          <ProtectedRoute
-            // logged in shows InfoPage else shows LoginPage
-            exact
-            path="/info"
-          >
-            <InfoPage />
-          </ProtectedRoute>
-
-          <Route
-            exact
-            path="/login"
-          >
-            {user.id ?
-              // If the user is already logged in, 
-              // redirect to the /user page
-              <Redirect to="/admin" />
-              :
-              // Otherwise, show the login page
-              <LoginPage />
-            }
-          </Route>
-
-          <Route
-            exact
-            path="/registration"
-          >
-            {user.id ?
-              // If the user is already logged in, 
-              // redirect them to the /user page
-              <Redirect to="/admin" />
-              :
-              // Otherwise, show the registration page
-              <RegisterPage />
-            }
-          </Route>
 
           <Route
             exact
             path="/home"
           >
-            {user.id ?
-              // If the user is already logged in, 
-              // redirect them to the /user page
-              <Redirect to="/user" />
-              :
-              // Otherwise, show the Landing page
               <LandingPage />
-            }
+          </Route>
+
+          <Route
+           exact
+           path="/admin">
+            <Admin/>
           </Route>
 
           {/* If none of the other routes matched, we will show a 404. */}
