+++ src/components/LoginPage/LoginPage.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import LoginForm from '../LoginForm/LoginForm';
 import { useHistory } from 'react-router-dom';
+import Nav from '../Nav/Nav';
 
 function LoginPage() {
   const history = useHistory();
@@ -8,7 +9,6 @@ function LoginPage() {
   return (
     <div>
       <LoginForm />
-
       <center>
         <button
           type="button"
