+++ src/index.js
@@ -5,6 +5,7 @@ import { Provider } from 'react-redux';
 import store from './redux/store';
 
 import App from './components/App/App';
+import Admin from './components/Admin/Admin';
 
 ReactDOM.render(
   <Provider store={store}>
